todays updates up to imports dimensions and panels

+++ webeditr/editor/views/sheet_editor.py
@@ -136,50 +136,6 @@ def add_stylesheet_attribute(request):
 
 
 @never_cache
-def add_script_function(request):
-    try:
-        rdict = dict(request.POST)
-        scriptsheet_name = escape(rdict['scriptsheet_name'][0])
-        func_name = escape(rdict['scriptsheet_name'][0])
-        func_script = rdict['scriptsheet_script'][0]
-        func_description = escape(rdict['scriptsheet_script'][0])
-        sheet = ScriptSheet(name=scriptsheet_name)
-        if sheet.count() > 0:
-            sheet = sheet.first()
-            func, created = ScriptFunc.objects.get_or_create(name=func_name, description=func_description, func=func_script)
-            func.save()
-            script_func, created = ScriptScriptSheet.objects.get_or_create(script_sheet_id=sheet.id, func_id=func.id)
-            script_func.save()
-        else:
-            return JsonResponse({'success': False, 'msg': 'Scriptsheet Not Found'})
-    except Exception as e:
-        print(traceback.format_exc()) # send to elastic apm
-        return JsonResponse({'success': False, 'msg': 'Internal Error'})
-
-
-@never_cache
-def remove_scriptsheet_function(request):
-    try:
-        rdict = dict(request.POST)
-        func_name = escape(rdict['function_name'][0])
-        script_sheet_name = escape(rdict['scriptsheet_name'][0])
-        script = ScriptSheet.object.filter(name=script_sheet_name)
-        if script.count() > 0:
-            func = ScriptFunc.objects.filter(name=func_name)
-            if func.count() > 0:
-                func = func.first()
-                script = ScriptScriptSheet.object.filters(script_sheet_id=script.first().id, func_id=func.id)
-                if script.count() > 0:
-                    script.delete()
-                script = ScriptScriptSheet.objects.filter(func_id=func.id)
-                if script.count() is 0 and func.count() > 0:
-                    func.delete()
-    except Exception as e:
-        print(traceback.format_exc()) # send to elastic apm
-        return JsonResponse({'success': False, 'msg': 'Internal Error'})
-
-
-@never_cache
 def remove_stylesheet_attribute(request):
     try:
         rdict = dict(request.POST)
@@ -188,13 +144,24 @@ def remove_stylesheet_attribute(request):
         attribute_name = escape(rdict['attribute_name'][0])
         style_sheet = StyleSheet.objects.filter(name=stylesheet_name)
         if style_sheet.count() > 0:
-            attr = Classes.objects.filter(name=attribute_name)
+            attr = Classes.objects.filter(name=class_name)
             if attr.count() > 0:
-                style_sheet_attr = ClassesStylesheet.objects.filter(style_sheet_id=style_sheet.first().id, classes_id=attr.first().id)
-                if style_sheet_attr.count() > 0:
-                    style_sheet_attr.delete()
+                classes = attr.first()
+                attrs = classes.attributes
+                if attrs.get(attribute_name, None) is not None:
+                    classes.attributes.pop(attribute_name)
+                    classes.save()
+                    if len(classes.attributes.keys()) is 0:
+                        style_sheet_attr = ClassesStylesheet.objects.filter(style_sheet_id=style_sheet.first().id, classes_id=attr.first().id)
+                        if style_sheet_attr.count() > 0:
+                            for sheet in style_sheet_attr:
+                                style_sheet_attr.delete()
+                        classes.delete()
+                    return JsonResponse({'success': True, 'class_id': classes.id})
+                else:
+                    return JsonResponse({'success': False, 'msg': 'Attribute Not Found'})
             else:
-                return JsonResponse({'success': False, 'msg': 'ATtribute Not Found'})
+                return JsonResponse({'success': False, 'msg': 'Class Not Found'})
         else:
             return JsonResponse({'success': False, 'msg': 'Stylesheet Not Found'})
     except Exception as e:
